chore: Rename ShowPerkAddonStore to VisualizationStore
